force plaintext for PGP/Inline scheme
